new file structure, redux implementation finished
