+++ projects/jammming/src/actions/types.js
@@ -1,0 +1,12 @@
+export const ADD_TRACK = 'ADD_TRACK'
+export const REMOVE_TRACK = 'REMOVE_TRACK'
+export const REQUEST_SEARCH = 'REQUEST_SEARCH'
+export const RECEIVE_SEARCH = 'RECEIVE_SEARCH'
+export const SAVE_NEEDED = 'SAVE_NEEDED'
+export const SAVE_FINISHED = 'SAVE_FINISHED'
+export const UPDATE_PLAYLIST_NAME = 'UPDATE_PLAYLIST_NAME'
+export const PLAYLISTS_REQUEST = 'PLAYLISTS_REQUEST'
+export const PLAYLISTS_RECIEVED = 'PLAYLISTS_RECEIVED'
+export const PLAYLIST_TRACKS_REQUEST = 'PLAYLIST_TRACKS_REQUEST'
+export const PLAYLIST_TRACKS_RECEIVED = 'PLAYLIST_TRACKS_RECEIVED'
+export const AUTH_RECEIVED = 'AUTH_RECEIVED'
