+++ projects/jammming/src/components/App/index.js
@@ -1,22 +1,18 @@
 import { connect } from 'react-redux';
 import App from './App';
-import { getAccessToken } from '../../actions/actions';
+import { getAccessToken } from '../../actions';
 
 
 const mapStateToProps = state => ({ tokenGranted: state.token.tokenGranted })
 
 const mapDispatchToProps = dispatch => {
-    return {
-      getAccessToken: () => {
-        dispatch(getAccessToken())
-      }
+  return {
+    getAccessToken: () => {
+      dispatch(getAccessToken())
     }
   }
+}
 
-//const SearchBarContainer = connect(mapStateToProps, 
- //   mapDispatchToProps
-//)(SearchBar);
-
-export default connect(mapStateToProps, 
-    mapDispatchToProps
+export default connect(mapStateToProps,
+  mapDispatchToProps
 )(App);
