--- projects/jammming/src/components/TrackList/TrackList.css
@@ -1,3 +1,0 @@
-.TrackList {
-    width: 100%;
-  }
