+++ projects/jammming/src/components/tracklistComponents/Tracklist/tracklist.css
@@ -1,0 +1,3 @@
+.Tracklist {
+    width: 100%;
+  }
